+++ src/functions/quota-collector/main.py
@@ -11,6 +11,7 @@ if __package__ in {None, ''}:
 from modules.qmchecks.ec2.ec2 import get_current_quotastatus_ec2
 from modules.qmchecks.ebs import get_current_quotastatus_ebs
 from modules.qmchecks.ec2_ipam import get_current_quotastatus_ec2_ipam
+from modules.qmchecks.customer_profiles import get_current_quotastatus_customer_profiles
 from modules.qmchecks.vpc.vpc import get_current_quotastatus_vpc
 from modules.qmchecks.lambda_checks.lambda_checks import get_current_quotastatus_lambda
 from modules.qmchecks.account_services import get_current_quotastatus_account_services
@@ -225,6 +226,7 @@ def lambda_handler(event, context):
         entries.extend(collector(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_ebs(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_ec2_ipam(ctx=ctx, skip=official))
+    entries.extend(get_current_quotastatus_customer_profiles(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_account_services(ctx, skip=official))
     entries.extend(get_current_quotastatus_elb(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_stepfunctions(ctx=ctx, skip=official))